Move Login view to the authentication module, refactor imports
