+++ frontend/src/authentication/contexts/CognitoAuthContext.js
@@ -1,7 +1,7 @@
 import Amplify, { Auth } from 'aws-amplify';
 import PropTypes from 'prop-types';
 import { createContext, useEffect, useReducer } from 'react';
-import { SET_ERROR } from '../../globalErrors/errorReducer';
+import { SET_ERROR } from 'globalErrors';
 
 Amplify.configure({
   Auth: {
