+++ frontend/src/routes.js
@@ -12,7 +12,7 @@ const Loadable = (Component) => (props) =>
 
 // Authentication pages
 const Login = Loadable(
-  lazy(() => import('./modules/Misc/views/authentication/Login'))
+  lazy(() => import('./authentication/views/Login'))
 );
 
 // Error pages
